+++ NEWS.md
@@ -3,6 +3,7 @@
 ### Version 0.9.14.9000 June 10, 2024
 
 - Updated buildAlignments() to build DPA2 and DPB2 cDNA alignments in releases prior to 3.54.0.
+- Updated buildAlignments() to build DRB1 gDNA alignments in release 3.48.0.
 
 ### Version 0.9.13.9000 June 4, 2024